Move ExternalUser model to a file
